requirements.txt
requests
python-daemon
pyyaml
websocket-client
slackclient
slacker
